Fixed unit test, moved docker stats to separate package to refactor and remove duplicate code
